added recent storage to collector
basically i implemented a file I/O for collector to store and access
ecents. puteCents() accepts a single string of ecents separated by
spaces, which it splits up into a list and stores in the file
ecents.txt. geteCents() accesses the first ecent in the list and then
removes it from said list